# swagger/mahakam.yaml
swagger: "2.0"
info:
  description: API definitions for Mahakam Cloud Platform
  title: Mahakam API
  version: 1.0.0
consumes:
- application/json
produces:
- application/json
schemes:
- http
basePath: /v1
paths:
  /clusters:
    get:
      tags:
        - clusters
      operationId: getClusters
      parameters:
        - in: query
          type: string
          name: owner
          description: Owner of clusters
      responses:
        200:
          description: list available clusters
          schema:
            type: array
            items:
              $ref: "#/definitions/cluster"
        default:
          description: generic error response
          schema:
            $ref: "#/definitions/error"
    post:
      tags:
        - clusters
      operationId: createCluster
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/cluster"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/cluster"
        default:
          description: error
          schema:
            $ref: "#/definitions/error"
...
  /networks:
    get:
      tags:
        - networks
      operationId: getNetworks
      responses:
        200:
          description: list created networks
          schema:
            type: array
            items:
              $ref: "#/definitions/network"
        default:
          description: generic error response
          schema:
            $ref: "#/definitions/error"
    post:
      tags:
        - networks
      operationId: createNetwork
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/network"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/network"
        default:
          description: error
          schema:
            $ref: "#/definitions/error"
...
definitions:
  cluster:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
        readOnly: true
      nodeSize:
        type: string
        default: "small"
        enum:
          - "extrasmall"
          - "small"
          - "medium"
          - "large"
          - "extralarge"
      name:
        type: string
        minLength: 1
      # TODO(giri): make owner a required field once we have proper auth
      owner:
        type: string
        minLength: 1
      numNodes:
        type: integer
        minimum: 1
        maximum: 10
      nodeFailures:
        type: array
        items:
          type: string
      componentFailures:
        type: array
        items:
          type: string
      podFailures:
        type: array
        items:
          type: string
      status:
        type: string
  network:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
        readOnly: true
      name:
        type: string
      networkName:
        type: string
      networkCIDR:
        type: string
      gateway:
        type: string
      nameserver:
        type: string
      dhcp:
        type: string
...
  error:
    type: object
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string